Fix: Edit MyPage icon code
