+++ src/icons/MyPage/HeartIcon.tsx
@@ -1,0 +1,18 @@
+import { SvgIcon, SvgIconProps } from '@mui/material'
+
+const HeartIcon = (props: SvgIconProps) => {
+  return (
+    <SvgIcon {...props}>
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        viewBox="0 0 12 10"
+        fill="none"
+        stroke="currentColor"
+      >
+        <path d="M11.7097 2.0493C11.5235 1.64828 11.255 1.28487 10.9193 0.979431C10.5833 0.673079 10.1872 0.429625 9.75251 0.262308C9.30173 0.0881223 8.81824 -0.001036 8.33012 9.0821e-06C7.64532 9.0821e-06 6.97719 0.174444 6.39657 0.503933C6.25766 0.582752 6.1257 0.669324 6.00069 0.763648C5.87567 0.669324 5.74371 0.582752 5.60481 0.503933C5.02419 0.174444 4.35605 9.0821e-06 3.67125 9.0821e-06C3.17814 9.0821e-06 2.70031 0.0878729 2.24887 0.262308C1.81271 0.430283 1.41961 0.671908 1.08207 0.979431C0.745914 1.28453 0.477385 1.64802 0.2917 2.0493C0.0986224 2.46665 0 2.90985 0 3.36597C0 3.79624 0.0944553 4.2446 0.281977 4.70072C0.438939 5.08189 0.663965 5.47728 0.951498 5.87654C1.40711 6.50839 2.03357 7.16736 2.81143 7.83539C4.10047 8.94273 5.377 9.70766 5.43118 9.73867L5.76038 9.93507C5.90623 10.0216 6.09375 10.0216 6.2396 9.93507L6.56881 9.73867C6.62298 9.70637 7.89813 8.94273 9.18855 7.83539C9.96642 7.16736 10.5929 6.50839 11.0485 5.87654C11.336 5.47728 11.5624 5.08189 11.718 4.70072C11.9055 4.2446 12 3.79624 12 3.36597C12.0014 2.90985 11.9028 2.46665 11.7097 2.0493Z" />
+      </svg>
+    </SvgIcon>
+  )
+}
+
+export default HeartIcon
