+++ src/icons/MyPage/SettingIcon.tsx
@@ -1,0 +1,18 @@
+import { SvgIcon, SvgIconProps } from '@mui/material'
+
+const SettingIcon = (props: SvgIconProps) => {
+  return (
+    <SvgIcon {...props}>
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        viewBox="0 0 12 12"
+        fill="none"
+        stroke="currentColor"
+      >
+        <path d="M6.0023 4.36619C5.57892 4.36619 5.18244 4.52177 4.88225 4.8061C4.58347 5.09043 4.4178 5.46596 4.4178 5.86697C4.4178 6.26798 4.58347 6.64351 4.88225 6.92784C5.18244 7.21083 5.57892 7.36774 6.0023 7.36774C6.42569 7.36774 6.82217 7.21083 7.12236 6.92784C7.42114 6.64351 7.58681 6.26798 7.58681 5.86697C7.58681 5.46596 7.42114 5.09043 7.12236 4.8061C6.97574 4.66616 6.80122 4.55521 6.60894 4.47969C6.41666 4.40417 6.21046 4.36559 6.0023 4.36619ZM11.8405 7.52466L10.9144 6.77494C10.9583 6.52012 10.981 6.25993 10.981 6.00109C10.981 5.74224 10.9583 5.48071 10.9144 5.22723L11.8405 4.47751C11.9104 4.42079 11.9605 4.34524 11.984 4.26091C12.0075 4.17658 12.0034 4.08747 11.9722 4.00542L11.9594 3.97055C11.7046 3.29552 11.3227 2.66982 10.8323 2.12375L10.8068 2.09559C10.7472 2.02927 10.6679 1.98159 10.5792 1.95885C10.4904 1.93611 10.3965 1.93936 10.3098 1.96817L9.15999 2.35577C8.73519 2.02584 8.26224 1.76566 7.74965 1.5846L7.52734 0.445941C7.51056 0.360162 7.46663 0.281248 7.40138 0.219681C7.33612 0.158115 7.25264 0.116812 7.16201 0.101259L7.12378 0.0945528C6.38746 -0.0315176 5.61149 -0.0315176 4.87517 0.0945528L4.83693 0.101259C4.74631 0.116812 4.66282 0.158115 4.59756 0.219681C4.53231 0.281248 4.48838 0.360162 4.47161 0.445941L4.24788 1.58996C3.74014 1.77246 3.26715 2.03202 2.84745 2.35846L1.68916 1.96817C1.60246 1.93913 1.50847 1.93576 1.41969 1.95852C1.33091 1.98128 1.25155 2.02908 1.19215 2.09559L1.16666 2.12375C0.677137 2.6704 0.295353 3.29595 0.0395221 3.97055L0.0267781 4.00542C-0.0369419 4.17307 0.0154501 4.36083 0.158466 4.47751L1.09586 5.23528C1.05196 5.48742 1.03072 5.74492 1.03072 5.99975C1.03072 6.25725 1.05196 6.51476 1.09586 6.76422L0.161298 7.52198C0.0913417 7.5787 0.0412729 7.65425 0.0177496 7.73858C-0.00577364 7.82291 -0.00163684 7.91202 0.0296101 7.99407L0.0423541 8.02894C0.29865 8.70355 0.676723 9.3272 1.16949 9.87574L1.19498 9.90391C1.25453 9.97022 1.33389 10.0179 1.42262 10.0406C1.51134 10.0634 1.60526 10.0601 1.692 10.0313L2.85028 9.64104C3.27225 9.96962 3.74237 10.2298 4.25071 10.4095L4.47444 11.5536C4.49121 11.6393 4.53514 11.7182 4.6004 11.7798C4.66565 11.8414 4.74914 11.8827 4.83977 11.8982L4.878 11.9049C5.62156 12.0317 6.38304 12.0317 7.12661 11.9049L7.16484 11.8982C7.25547 11.8827 7.33896 11.8414 7.40421 11.7798C7.46946 11.7182 7.5134 11.6393 7.53017 11.5536L7.75248 10.4149C8.26507 10.2325 8.73802 9.97365 9.16282 9.64372L10.3126 10.0313C10.3993 10.0604 10.4933 10.0637 10.5821 10.041C10.6709 10.0182 10.7502 9.97041 10.8096 9.90391L10.8351 9.87574C11.3279 9.32452 11.706 8.70355 11.9623 8.02894L11.975 7.99407C12.0359 7.82777 11.9835 7.64134 11.8405 7.52466ZM6.0023 8.22475C4.62737 8.22475 3.51297 7.16925 3.51297 5.86697C3.51297 4.56469 4.62737 3.50918 6.0023 3.50918C7.37724 3.50918 8.49163 4.56469 8.49163 5.86697C8.49163 7.16925 7.37724 8.22475 6.0023 8.22475Z" />
+      </svg>
+    </SvgIcon>
+  )
+}
+
+export default SettingIcon
