+++ src/icons/MyPage/WhaleIcon.tsx
@@ -5,15 +5,12 @@ const WhaleIcon = (props: SvgIconProps) => {
     <SvgIcon {...props}>
       <svg
         xmlns="http://www.w3.org/2000/svg"
-        width="16"
-        height="16"
         viewBox="0 0 16 16"
         fill="none"
+        stroke="currentColor"
       >
         <path
           d="M3.36224 6.66369C3.42068 6.26013 3.54219 5.95775 3.68758 5.72932C3.89297 5.40661 4.16687 5.19879 4.45146 5.06279C4.65825 4.96398 4.86856 4.90432 5.05524 4.86887C5.00394 5.07311 4.92603 5.30759 4.80866 5.53893C4.65294 5.84587 4.43412 6.13516 4.12658 6.34823C3.9264 6.48692 3.67706 6.60148 3.36224 6.66369ZM1.37571 6.10643C0.99898 5.79792 0.691722 5.38542 0.56394 4.86441C0.757706 4.89889 0.97936 4.95936 1.19716 5.06326C1.48203 5.19914 1.75621 5.40692 1.96181 5.72957C2.10076 5.94763 2.21791 6.23309 2.27913 6.60967C1.99436 6.51346 1.67157 6.34872 1.37571 6.10643ZM4.0236 10.0295C4.2484 10.2259 4.50176 10.3896 4.78029 10.4761C5.24759 10.6213 5.73103 10.531 6.15037 10.1603C6.58343 9.77749 6.95299 9.18528 7.29742 8.58955C7.39544 8.42001 7.49248 8.24821 7.59014 8.0753C7.85145 7.61267 8.11721 7.14215 8.41814 6.6861C9.24421 5.43422 10.2045 4.5 11.6368 4.5C12.4935 4.5 13.1449 4.68862 13.6459 5.0067C14.1466 5.32457 14.5328 5.79445 14.8219 6.41616C15.307 7.45918 15.1326 8.54849 14.5201 9.53581C13.9021 10.532 12.8535 11.389 11.6872 11.8869C10.7397 12.2915 9.72014 12.5 8.68991 12.5H8.37924C7.83512 12.5 7.3307 12.4577 6.87357 12.3015C5.6266 11.8756 4.63746 11.1007 4.0236 10.0295Z"
-          fill="#42444C"
-          stroke="#42444C"
           strokeLinejoin="round"
         />
       </svg>
