+++ src/icons/MyPage/index.ts
@@ -1,0 +1,5 @@
+export { default as HeartIcon } from './HeartIcon'
+export { default as MessageIcon } from './MessageIcon'
+export { default as SettingIcon } from './SettingIcon'
+export { default as SilhouetteIcon } from './SilhouetteIcon'
+export { default as WhaleIcon } from './WhaleIcon'
