+++ src/icons/index.ts
@@ -3,9 +3,3 @@ export { default as CloseIcon } from './CloseIcon'
 export { default as PlusIcon } from './PlusIcon'
 export { default as SearchIcon } from './SearchIcon'
 export { default as SendIcon } from './SendIcon'
-
-export { default as MyPageHeartIcon } from './MyPageHeartIcon'
-export { default as FilledMessageIcon } from './FilledMessageIcon'
-export { default as SettingIcon } from './SettingIcon'
-export { default as SilhouetteIcon } from './SilhouetteIcon'
-export { default as WhaleIcon } from './WhaleIcon'
